Fix build errors on CentOS 7 and potential logic error with int, size_t

+++ src/parametric.cpp
@@ -530,14 +530,14 @@ bool ParametricViewer::ExportFromMacro(wxString path, bool asExcel) {
 	return true;
 }
 
-bool ParametricViewer::SetInputFromMacro(wxString varName, size_t index, wxString val) {
+bool ParametricViewer::SetInputFromMacro(wxString varName, int index, wxString val) {
 	int c = m_grid_data->GetColumnForName(varName);
 	if (c == wxNOT_FOUND)
 		return false;
-	if (index < 0 || index >= (int)m_grid_data->GetNumberRows())
+	if (index < 0 || index >= m_grid_data->GetNumberRows())
 		return false;
-	m_grid_data->SetValue((int)index, c, val);
-	if (m_grid_data->GetValue((int)index, c) == val) {
+	m_grid_data->SetValue(index, c, val);
+	if (m_grid_data->GetValue(index, c) == val) {
 		m_grid->SetTable(m_grid_data);
 		UpdateGrid();
 		return true;
